Make version name indicate in-development versions
