+++ CONTRIBUTING.md
@@ -92,8 +92,9 @@ For a new file, add this license header to the top of the file:
 3. Run `npm ci` to install all dependencies for building from [package-lock.json](package-lock.json).
 4. Update [CHANGELOG.md](CHANGELOG.md) with the update changelog for the new version.
 5. Make a new git commit and tag with `git commit -S`,  `git tag VERSION -a`, add the new version name as the annotation, then run `git push origin VERSION`.
-6. Run `npm run clean`, `npm run webpack:build:firefox`, then `npm run package`. Upload the resulting package to [AMO](https://addons.mozilla.org/en-US/developers/addons) to have the extension signed. Upload the result into GitHub releases then update the update server to point to the new release.
-7. Run `npm run clean`, `npm run webpack:build:chromium`, then `npm run package`. Upload the resulting package to [CWS](https://chrome.google.com/webstore/developer/dashboard) to have the new version signed and released.
+6. Run `npm run libraries` to generate third-party library attributions.
+7. Run `npm run clean`, `SASPES_OFFICIAL_RELEASE=true npm run webpack:build:firefox`, then `npm run package`. Upload the resulting package to [AMO](https://addons.mozilla.org/en-US/developers/addons) to have the extension signed. Upload the result into GitHub releases then update the update server to point to the new release.
+8. Run `npm run clean`, `SASPES_OFFICIAL_RELEASE=true npm run webpack:build:chromium`, then `npm run package`. Upload the resulting package to [CWS](https://chrome.google.com/webstore/developer/dashboard) to have the new version signed and released.
 
 ### Developer Certificate of Origin
 ```
