+++ package.json
@@ -35,6 +35,7 @@
     "ava": "^2.4.0",
     "babel-loader": "^8.0.6",
     "chromedriver": "^78.0.1",
+    "command-exists": "^1.2.8",
     "copy-webpack-plugin": "^5.1.1",
     "css-loader": "^3.2.0",
     "eslint": "^6.5.1",
