+++ src/ui/options.html
@@ -1,18 +1,19 @@
 <!DOCTYPE html>
 <html>
 <head>
-    
+
 </head>
 <body>
-    <form>   
+    <h3>Version: <span id="version"></span></h3>
+    <form>
         <label><input type="checkbox" id="analytics" />Enable Analytics</label>
-        <br />   
+        <br />
         <label><input type="checkbox" id="percent-mp" />Load final percent on overall grade page</label>
         <br />
         <!--<label><input type="checkbox" id="save-grades" />Save last seen grades (requires previous option to work)</label>-->
     </form>
     <br />
-    
+
     <footer>
         <br />
         Last Seen Grades has been temporarily removed to be improved upon.<br />
