+++ src/ui/options.js
@@ -1,13 +1,13 @@
 /**
- * 
+ *
  * @copyright Copyright (c) 2018-2019 Gary Kim <gary@garykim.dev>
- * 
+ *
  * @author Gary Kim <gary@garykim.dev>
- * 
+ *
  * SAS Powerschool Enhancement Suite - A browser extension to improve the experience of SAS Powerschool.
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as 
+ * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, version 3.
  *
  * This program is distributed in the hope that it will be useful,
@@ -17,7 +17,7 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- * 
+ *
  */
 
 'use strict';
@@ -37,6 +37,7 @@ function main() {
         analyticsid = returned.id;
     });
     document.getElementById("third-party-libraries").href = browser.extension.getURL("web_accessible_resources/libraries.txt");
+    document.getElementById("version").innerText = SASPES_VERSION_NAME;
     document.getElementById("analytics").addEventListener("click", function() {
         let value = document.getElementById("analytics").checked;
         if(!value)  {
