+++ tests/extensionInfo.js
@@ -1,7 +1,7 @@
 /**
- * 
- * @copyright Copyright (c) 2019 Gary Kim <gary@garykim.dev>
- * 
+ *
+ * @copyright Copyright (c) 2019-2020 Gary Kim <gary@garykim.dev>
+ *
  * @author Gary Kim <gary@garykim.dev>
  *
  * @license GNU AGPL version 3
@@ -9,7 +9,7 @@
  * SAS Powerschool Enhancement Suite - A browser extension to improve the experience of SAS Powerschool.
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as 
+ * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, version 3.
  *
  * This program is distributed in the hope that it will be useful,
@@ -19,10 +19,10 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- * 
+ *
  */
 
-const package = require('../package.json');
+const version_name = require('../webpack-helpers').versionName();
 
 module.exports = {
     "Show Extension Info": browser => {
@@ -33,7 +33,7 @@ module.exports = {
     "Extension Version Correct": browser => {
         browser
             .assert.containsText('#saspes-info', 'SAS Powerschool Enhancement Suite')
-            .assert.containsText("#saspes-info", `Version: ${package.version_name}`)
+            .assert.containsText("#saspes-info", `Version: ${version_name}`)
     },
     "Extension Info Remember Open State": browser => {
         browser
